fix: 빌드 버그 수정
