+++ src/app/(pages)/mypage/layout.tsx
@@ -1,3 +1,4 @@
+import { Suspense } from "react";
 import FilterBar from "./components/FilterBar";
 
 interface MypageLayoutProps {
@@ -8,8 +9,10 @@ export default function MypageLayout({ children }: MypageLayoutProps) {
   return (
     <div className="flex min-h-screen flex-col">
       <div className="px-6 md:px-[72px] lg:px-[220px]">
-        <FilterBar />
-        {children}
+        <Suspense fallback={<div>로딩 중...</div>}>
+          <FilterBar />
+          {children}
+        </Suspense>
       </div>
     </div>
   );
