+++ src/app/(pages)/mypage/page.tsx
@@ -1,19 +1,7 @@
 "use client";
 
-import { useSearchParams } from "next/navigation";
-import PostsSection from "./components/sections/PostsSection";
-import CommentsSection from "./components/sections/CommentsSection";
-import ScrapsSection from "./components/sections/ScrapsSection";
+import MyPageContent from "./components/MyPageContent";
 
 export default function MyPage() {
-  const searchParams = useSearchParams();
-  const currentTab = searchParams.get("tab") || "posts";
-
-  const TabContent = {
-    posts: <PostsSection />,
-    comments: <CommentsSection />,
-    scrap: <ScrapsSection />,
-  }[currentTab];
-
-  return TabContent;
+  return <MyPageContent />;
 }
